requirements.txt
Flask==3.0.3
requests==2.32.3
openai==1.40.0
python-dotenv==1.0.1
PyPDF2==3.0.1
PyMuPDF==1.24.9
pytesseract==0.3.13
nougat-ocr==0.1.0
google-cloud-texttospeech==2.7.0
certifi==2024.7.4
